requirements.txt
beautifulsoup4==4.13.3
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
